+++ circleci/install-python-2.5.6.sh
@@ -1,7 +1,7 @@
 set -x
 set -e
 
-BUILDROOT=$HOME/pg8000
+BUILDROOT=$HOME/aiopg8000
 
 if [[ ! -e py-2.5.6/bin/python2.5 ]]; then
     wget https://www.python.org/ftp/python/2.5.6/Python-2.5.6.tar.bz2